+++ SwiftGui/SwiftGui/Attributes/Dir.swift
@@ -1,0 +1,18 @@
+//
+//  Dir.swift
+//  SwiftGui
+//
+//  Created by Erick Jung on 07/05/2020.
+//  Copyright © 2020 Erick Jung. All rights reserved.
+//
+
+import Foundation
+
+public enum ImGuiDir: Int32
+{
+    case none  = -1
+    case left  = 0
+    case right = 1
+    case up    = 2
+    case down  = 3
+}
