+++ SwiftGui/SwiftGui/Attributes/Edge.swift
@@ -1,0 +1,18 @@
+//
+//  Edge.swift
+//  SwiftGui
+//
+//  Created by Erick Jung on 07/05/2020.
+//  Copyright © 2020 Erick Jung. All rights reserved.
+//
+
+import Foundation
+
+public enum GUIEdge
+{
+    case all
+    case top
+    case leading
+    case bottom
+    case trailing
+}
